requirements.txt
black==21.9b0
blinker==1.8.2
click==8.1.7
Flask==2.0.3
Flask-WTF==1.0.1
importlib_metadata==7.1.0
isort==5.10.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
numpy==1.24.4
pandas==1.2.2
pathspec==0.12.1
platformdirs==4.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.5.15
scikit-surprise==1.1.1
scipy==1.10.1
six==1.16.0
SQLAlchemy==1.3.18
tomli==1.2.3
typing_extensions==4.12.2
Werkzeug==2.0.3
WTForms==3.1.2
zipp==3.19.2
